+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- 🪟 Windows XP support! (windows-legacy build)
+
 ## [v5.1.8] - 2026-02-14
 
 ### Fixed